\name{palmdates}
\docType{data}
\alias{palmdates}

\title{Spectrometry and composition chemical of Spanish and Arabian palm dates}
\description{

 A data set with scores of 21 dates on spectrometry and concentration measurements of the substances that better define the quality of the dates: fibre, sorbitol, fructose, glucose and myo-inositol. The first 11 dates are Spanish (from Elche, Alicante) and the last 10 are from other countries, mainly Arabian.
}
\usage{palmdates}
\format{
  A data frame with 2 elements:
\describe{
  \item{conc}{a data frame with 5 columns: fibre, sorbitol, fructose, glucose and myo-inositol.}
  \item{spectra}{a data frame with 2050 columns.}
  }
}

\references{

Abdrabo, S.S., Gras, L., Grindlay, G. and Mora, J. (2021) Evaluation of Fourier Transform-Raman Spectroscopy for palm dates characterization. Journal of food composition and analysis. Submitted.
}

\author{Maria Jose Nueda, \email{mj.nueda@ua.es}}


\keyword{datasets}

