\name{predict.lpda}
\alias{predict.lpda}

\title{Predict method for lpda classification}

\description{
Predict method for lpda classification
}

\usage{
\method{predict}{lpda}(object, datatest = object$data,...)
}

\arguments{
\item{object}{Object of class inheriting from "lpda"}

\item{datatest}{An optional data to predict their class. If omitted, the original data is used.}

\item{...}{Other arguments passed.}
}

\author{Maria Jose Nueda, \email{mj.nueda@ua.es}}

\seealso{ \code{\link{lpda}}}

\keyword{classification}
