\name{RNAseq}
\docType{data}
\alias{RNAseq}

\title{Simulated RNA-Seq dataset example}
\description{

  A simulated RNA-Seq dataset example.

}
\usage{RNAseq}
\format{
  A data frame with 600 variables (in columns) and 60 samples (rows).

}

\details{
  This dataset is a RNA-Seq simulated example.
  It has been simulated as Negative Binomial distributed and transformed to rpkm  (Reads per kilo base per million mapped reads).
  It contains  600 genes (in columns) and 60 samples (rows), 30 of each one of the experimental groups.
  First 30 samples are from first group and the remaining samples from the second one.
}

 \author{Maria Jose Nueda, \email{mj.nueda@ua.es}}


\keyword{datasets}

