\name{stand2}
\alias{stand2}

\title{stand2 center and scale a data matrix with the parameters of another one}

\description{
\code{stand2} center and scale a data matrix with the parameters of another one
}

\usage{
stand2(X, X2)
}
\arguments{
\item{X}{the data matrix from which mean and standard deviation is computed}

\item{X2}{the data matrix to center and scale}
}
