\name{RNAseq}
\docType{data}
\alias{RNAseq}

\title{Simulated RNA-Seq Time Series 3-dimensional dataset example}
\description{

  A simulated RNA-Seq dataset example.

}
\usage{RNAseq}
\format{
  An array data frame with 20 samples (1st dimension), 600 genes (2nd dimension) and 4 time-points (3rd dimension).

}

\details{
  This dataset is a RNA-Seq simulated example.
  It has been simulated as Negative Binomial distributed and transformed to rpkm  (Reads per kilo base per million mapped reads).
  It contains gene expression from 600 genes measured to 60 samples through 4 time-points.
  First 10 samples are from first group and the remaining samples from the second one.
}

 \author{Maria Jose Nueda, \email{mj.nueda@ua.es}}


\keyword{datasets}

